Update scrape_sites.py
Adding the story function for The East African (Kenya). The site does appear to have a sitemap, but either the sitemap is broken or the link is out of date; consequently I have not written gen or collect functions. The sitemap allegedly resides at the following link: https://www.theeastafrican.co.ke/sitemap/sitemap-index.xml