Update scrape_sites.py
Adding story function for Bloomberg. 

I have not written the gen or collect functions for Bloomberg yet because of their html-download-prevention habits. 

I haven't tested this function because bloomberg is preventing me from downloading the html code. consequently, I can look at the code from the website and write a scraper based on what I think it should be, but I can't verify that the function works.